requirements.txt
asgiref==3.5.2
coverage==6.4.1
Django==4.0.5
django-appconf==1.0.5
django-filter==22.1
django-imagekit==4.1.0
djangorestframework==3.13.1
Markdown==3.3.7
pilkit==2.0
Pillow==9.1.1
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
